requirements.txt
annotated-types==0.7.0
bertopic==0.17.3
blinker==1.9.0
blis==1.2.1
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
emoji==2.14.1
fasttext-wheel==0.9.2
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
gensim==4.3.3
hdbscan==0.8.40
huggingface-hub==0.36.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.45.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.10.0
networkx==3.5
nltk==3.9.1
numba==0.62.1
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.3.1
preshed==3.0.10
protobuf==6.32.1
pybind11==3.0.1
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
Sastrawi==1.0.1
scikit-learn==1.7.2
scipy==1.13.1
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stanza==1.10.1
sympy==1.14.0
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
